UniHub/requirements.txt
Django
django-cors-headers
Pillow
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
psycopg2-binary
python-dotenv
celery==5.5.2
redis==6.0.0